requirements.txt
opencv-python
numpy
scikit-image
pyinstaller
tkeasyGUI
Pillow
